consumer/requirements.txt
fastapi
uvicorn
boto3
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary